// ui/src/app/core/enums/werewolves/werewolvesCharacterType.enum.ts
//   VILLAGER, // 村民
//   SEER, // 预言家
//   WITCH, // 女巫
//   HUNTER, // 猎人
//   KNIGHT, // 骑士
//   IDIOT, // 白痴
//   GUARD, // 守卫
//   WILD_CHILD, // 野孩子
//   BASTARD, // 混子
//   WEREWOLF, // 狼人
//   WEREWOLF_KING, // 狼王
//   WEREWOLF_WHITE_KING, // 白狼王
//   WEREWOLF_BEAUTY // 狼美人
export enum WerewolvesCharacterType {
  VILLAGER = "VILLAGER",
  SEER = "SEER",
  WITCH = "WITCH",
  HUNTER = "HUNTER",
  KNIGHT = "KNIGHT",
  IDIOT = "IDIOT",
  GUARD = "GUARD",
  WILD_CHILD = "WILD_CHILD",
  BASTARD = "BASTARD",
  WEREWOLF = "WEREWOLF",
  WEREWOLF_KING = "WEREWOLF_KING",
  WEREWOLF_WHITE_KING = "WEREWOLF_WHITE_KING",
  WEREWOLF_BEAUTY = "WEREWOLF_BEAUTY",
}